Started adding texturing system
Started adding the code to support texturing. First we need to load the asset and then, create a canvas for it, where the pixels will be loaded. We need this information in a buffer so we can get the pixel colors from the texel information after we linear interpolate the position of the pixel